Decision about how to handle svg and their styling - move the code in svg files to the html
There were two options: either have the svg code in the svg file, and import that as <object> to html.
With that way, we cannot have an external stylesheet to style the svg, so all the CSS has to be in svg files.
This would end up replicating a lot of CSS code, and since this repo has a lot of CSS code, I chose the other way...

...which is moving the svg code to the html, and therefore getting to modify it with external stylesheet.
This replicates the svg code (so both of the options disadvantage is replication), but it's better to replicate
the svg, which should change, and get to optimise and write CSS in one place.

Kept the old svg files still in repo, so that they're easy to find.