New decision about svg and styling - decided to get the svg as an img tag, and styles as inline styles in that file
All letters are now in one file, so this reduces the amount of style repetition. Also, this way the img is responsive
and looks nice even on mobile screen. The only downside is, that it is not possible to have a hover transition,
but one cannot have everything in life.